requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boto3==1.34.143
botocore==1.34.143
bravado==11.0.3
bravado-core==6.1.1
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.1.0
-e git+https://github.com/ahmadfaizalbh/Chatbot.git@8ef4fd5824d650f837aef7eac5e290f51cc9d463#egg=chatbotAI
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.1.5
cryptography==42.0.8
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
detect-secrets==1.5.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
ggshield==1.29.0
gitdb==4.0.11
gitdb2==4.0.2
GitPython==3.0.6
google-auth==2.31.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
griffe==0.36.9
grpcio==1.64.1
guardrails-ai==0.4.5
guardrails-api-client==0.2.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
kubernetes==30.1.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.3
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.84
language_data==1.2.0
litellm==1.41.25
llama-cloud==0.0.9
llama-index==0.10.56
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.12
llama-index-core==0.10.56
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
lxml==4.9.4
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mdurl==0.1.2
mistralai==0.4.2
mlx==0.16.0
mlx-lm==0.15.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
neptune==1.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.10
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
phonenumbers==8.13.40
pillow==10.4.0
pinecone-client==3.2.2
platformdirs==3.0.0
pluggy==1.5.0
posthog==3.5.0
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydash==7.0.7
pygitguardian==1.15.2
Pygments==2.18.0
PyJWT==2.8.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pytest==8.3.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rebuff @ git+https://github.com/protectai/rebuff@4d2fe064abf164e7381556d23e48e210080f8afa#subdirectory=python-sdk
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
rstr==3.2.2
s3transfer==0.10.2
safetensors==0.4.3
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.37.2
striprtf==0.0.26
swagger-spec-validator==3.0.4
sympy==1.12.1
tenacity==8.5.0
thinc==8.2.5
tiktoken==0.7.0
tldextract==5.1.2
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
truffleHog==2.2.1
truffleHogRegexes==0.0.7
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
webcolors==24.6.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
